arty: add csr for cpu_reset
This does not work. It's supposed to provide a CSR that controls the
reset of the softcore.

The CSR is correctly generated, and I can read/write to it. The value
holds which is unexpected. However, it does not cause the CPU to reset.

I tried inverting the signal, and that did not work.

Signed-off-by: Joel Stanley <joel@jms.id.au>